requirements.txt
gemini
streamlit
python-dotenv